move index pb to lance-index
