+++ rust/lance-index/build.rs
@@ -5,14 +5,7 @@ fn main() -> Result<()> {
 
     let mut prost_build = prost_build::Config::new();
     prost_build.protoc_arg("--experimental_allow_proto3_optional");
-    prost_build.compile_protos(
-        &[
-            "./protos/format.proto",
-            "./protos/index.proto",
-            "./protos/transaction.proto",
-        ],
-        &["./protos"],
-    )?;
+    prost_build.compile_protos(&["./protos/index.proto"], &["./protos"])?;
 
     Ok(())
 }
