+++ rust/lance-index/protos
@@ -1,0 +1,1 @@
+../../protos
