+++ rust/lance-index/src/lib.rs
@@ -15,3 +15,8 @@
 //! Lance secondary index library
 
 pub mod vector;
+
+pub mod pb {
+    #![allow(clippy::use_self)]
+    include!(concat!(env!("OUT_DIR"), "/lance.index.pb.rs"));
+}
